+++ data/examples/declaration/value/function/blank-lines-where-out.hs
@@ -1,0 +1,6 @@
+foo = x + y + z
+  where
+    x = 10
+
+    y = 11
+    z = 12