+++ data/examples/declaration/value/function/blank-lines-where.hs
@@ -1,0 +1,7 @@
+foo = x + y + z
+  where
+
+    x = 10
+
+    y = 11
+    z = 12